third_party/blink/web_tests/external/wpt/html/canvas/element/manual/imagebitmap/createImageBitmap-bounds-expected.txt

This is a testharness.js-based test.
PASS simple clip inside
FAIL clip outside negative assert_equals: expected 16 but got 8
FAIL clip outside positive assert_equals: expected 16 but got 8
FAIL clip inside using negative width and height assert_equals: expected 16 but got 8
Harness: the test ran to completion.
